+++ demo/local/topology.json
@@ -15,7 +15,7 @@
             "worker": "srv1",
             "type": "subproc",
             "working_dir": ".",
-            "cmd": "spout1.js",
+            "cmd": "spout_subproc.js",
             "args": [
                 "-a",
                 "action1"
@@ -40,7 +40,8 @@
             "name": "bolt1",
             "worker": "srv1",
             "working_dir": ".",
-            "cmd": "bolt1.js",
+            "type": "inproc",
+            "cmd": "bolt_inproc.js",
             "args": [
                 "-n",
                 "name1"
@@ -78,8 +79,8 @@
             "name": "bolt3",
             "worker": "srv1",
             "working_dir": ".",
-            "type": "inproc",
-            "cmd": "bolt_inproc.js",
+            "type": "subproc",
+            "cmd": "bolt_subproc.js",
             "args": [],
             "inputs": [
                 {
